関数切り出し
